+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2.1.1 under development
 
 - New #94: Add `getMessages()` method to `CategorySource` (@xepozz)
+- Bug #93: Throw exceptions on empty and not passed parameters, cast `null` to empty string in `SimpleMessageFormatter`
+  (@arogachev)
 
 ## 2.1.0 November 15, 2022